+++ util/date.js
@@ -1,5 +1,5 @@
 // LICENSE_CODE ZON ISC
-'use strict'; /*jslint node:true, browser:true*//* global BigInt */
+'use strict'; /*jslint node:true, browser:true*/
 (function(){
 var define;
 var is_node = typeof module=='object' && module.exports && module.children;
@@ -135,20 +135,6 @@ var monotonic_init = function(){
     }
 };
 monotonic_init();
-
-E.microsec = undefined;
-var microsec_init = function(){
-    var adjust;
-    if (is_node)
-    {
-        /* jshint -W119 */
-        var now_fn = function(){ return process.hrtime.bigint()/1000n; };
-        adjust = BigInt(Date.now())*1000n - now_fn();
-        E.microsec = function(){ return now_fn()+adjust; };
-        /* jshint +W119 */
-    }
-};
-microsec_init();
 
 E.str_to_dur = function(str, opt){
     opt = opt||{};